pyproject.toml
[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[project]
name = "rtx-trust"
version = "0.1.0"
description = "Real Tracker X: cross-ecosystem dependency trust scanner"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Andreas Fahl" }]
keywords = ["security", "supply-chain", "dependencies", "osv", "sbom"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "httpx==0.27.0",
    "rich==13.7.1",
    "packaging==24.1",
    "jinja2==3.1.4",
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.2",
    "mypy==1.11.1",
    "bandit==1.7.9",
    "semgrep==1.71.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.23.7",
    "pytest-cov==5.0.0",
    "hypothesis==6.112.2",
    "respx==0.20.2",
    "coverage==7.6.4",
    "types-PyYAML==6.0.12.20240917",
]
docs = [
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.14",
    "mkdocstrings[python]==0.25.1",
]
fuzz = [
    "atheris==2.3.0",
]

[project.urls]
Homepage = "https://github.com/afadesigns/rtx"
Documentation = "https://afadesigns.github.io/rtx"
Repository = "https://github.com/afadesigns/rtx"
Issues = "https://github.com/afadesigns/rtx/issues"
Changelog = "https://github.com/afadesigns/rtx/blob/main/CHANGELOG.md"

[project.scripts]
rtx = "rtx.cli:entrypoint"

[tool.hatch.build.targets.wheel]
packages = ["src/rtx"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "S", "ASYNC", "RUF"]
ignore = ["B905"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
plugins = []

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=rtx --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["rtx"]
omit = [
    "src/rtx/__main__.py",
]

[tool.bandit]
tests = ["B201", "B301", "B303", "B304", "B305", "B306", "B307", "B308", "B310", "B311", "B314", "B315", "B316", "B317", "B318", "B319", "B320", "B321", "B324", "B325", "B402", "B403", "B404", "B406", "B407", "B408", "B409", "B410", "B411", "B412", "B413", "B414", "B415", "B417", "B418", "B419", "B501", "B502", "B503", "B504", "B505", "B506", "B507", "B508", "B509"]

[tool.semgrep]
config = "p/ci"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["rtx"]

[tool.hatch.envs.default]
dependencies = [".[dev]"]

[tool.hatch.envs.docs]
dependencies = [".[docs]"]

[tool.hatch.envs.fuzz]
dependencies = [".[fuzz]"]
